+++ addons/actions/src/typings.d.ts
@@ -1,4 +1,3 @@
 // TODO: following packages need definition files or a TS migration
-declare module '@storybook/components';
 declare module 'global';
 declare module 'react-inspector';