+++ addons/knobs/src/components/Panel.js
@@ -21,7 +21,7 @@ import PropForm from './PropForm';
 
 const getTimestamp = () => +new Date();
 
-const DEFAULT_GROUP_ID = 'Other';
+export const DEFAULT_GROUP_ID = 'Other';
 
 const PanelWrapper = styled(({ children, className }) => (
   <ScrollArea horizontal vertical className={className}>
@@ -46,9 +46,9 @@ export default class KnobPanel extends PureComponent {
 
   componentDidMount() {
     this.mounted = true;
-    const { channel, api } = this.props;
-    channel.on(SET, this.setKnobs);
-    channel.on(SET_OPTIONS, this.setOptions);
+    const { api } = this.props;
+    api.on(SET, this.setKnobs);
+    api.on(SET_OPTIONS, this.setOptions);
 
     this.stopListeningOnStory = api.on(STORY_CHANGED, () => {
       if (this.mounted) {
@@ -60,9 +60,9 @@ export default class KnobPanel extends PureComponent {
 
   componentWillUnmount() {
     this.mounted = false;
-    const { channel } = this.props;
+    const { api } = this.props;
 
-    channel.removeListener(SET, this.setKnobs);
+    api.off(SET, this.setKnobs);
     this.stopListeningOnStory();
   }
 
@@ -72,7 +72,7 @@ export default class KnobPanel extends PureComponent {
 
   setKnobs = ({ knobs, timestamp }) => {
     const queryParams = {};
-    const { api, channel } = this.props;
+    const { api } = this.props;
 
     if (!this.options.timestamps || !timestamp || this.lastEdit <= timestamp) {
       Object.keys(knobs).forEach(name => {
@@ -80,15 +80,16 @@ export default class KnobPanel extends PureComponent {
         // For the first time, get values from the URL and set them.
         if (!this.loadedFromUrl) {
           const urlValue = api.getQueryParam(`knob-${name}`);
+
+          // If the knob value present in url
           if (urlValue !== undefined) {
-            // If the knob value present in url
-            knob.value = Types[knob.type].deserialize(urlValue);
-            channel.emit(CHANGE, knob);
+            const value = Types[knob.type].deserialize(urlValue);
+            knob.value = value;
+            queryParams[`knob-${name}`] = Types[knob.type].serialize(value);
+
+            api.emit(CHANGE, knob);
           }
         }
-
-        // set all knobsquery params to be deleted from URL
-        queryParams[`knob-${name}`] = null;
       });
 
       api.setQueryParams(queryParams);
@@ -99,9 +100,9 @@ export default class KnobPanel extends PureComponent {
   };
 
   reset = () => {
-    const { channel } = this.props;
+    const { api } = this.props;
 
-    channel.emit(RESET);
+    api.emit(RESET);
   };
 
   copy = () => {
@@ -119,9 +120,9 @@ export default class KnobPanel extends PureComponent {
   };
 
   emitChange = changedKnob => {
-    const { channel } = this.props;
+    const { api } = this.props;
 
-    channel.emit(CHANGE, changedKnob);
+    api.emit(CHANGE, changedKnob);
   };
 
   handleChange = changedKnob => {
@@ -138,9 +139,9 @@ export default class KnobPanel extends PureComponent {
   };
 
   handleClick = knob => {
-    const { channel } = this.props;
+    const { api } = this.props;
 
-    channel.emit(CLICK, knob);
+    api.emit(CLICK, knob);
   };
 
   render() {
@@ -196,9 +197,18 @@ export default class KnobPanel extends PureComponent {
       );
     }
 
-    const entries = Object.entries(groups);
-    // Always sort 'Other' (ungrouped) tab last without changing the remaining tabs
-    entries.sort((a, b) => (a[0] === 'Other' ? 1 : 0)); // eslint-disable-line no-unused-vars
+    // Always sort DEFAULT_GROUP_ID (ungrouped) tab last without changing the remaining tabs
+    const sortEntries = g => {
+      const unsortedKeys = Object.keys(g);
+      if (unsortedKeys.indexOf(DEFAULT_GROUP_ID) !== -1) {
+        const sortedKeys = unsortedKeys.filter(key => key !== DEFAULT_GROUP_ID);
+        sortedKeys.push(DEFAULT_GROUP_ID);
+        return sortedKeys.map(key => [key, g[key]]);
+      }
+      return Object.entries(g);
+    };
+
+    const entries = sortEntries(groups);
 
     return (
       <Fragment>
@@ -233,11 +243,6 @@ export default class KnobPanel extends PureComponent {
 KnobPanel.propTypes = {
   active: PropTypes.bool.isRequired,
   onReset: PropTypes.object, // eslint-disable-line
-  channel: PropTypes.shape({
-    emit: PropTypes.func,
-    on: PropTypes.func,
-    removeListener: PropTypes.func,
-  }).isRequired,
   api: PropTypes.shape({
     on: PropTypes.func,
     getQueryParam: PropTypes.func,