+++ addons/links/src/react/components/link.test.js
@@ -4,7 +4,7 @@ import addons from '@storybook/addons';
 
 import { SELECT_STORY } from '@storybook/core-events';
 import { mockChannel } from '../../preview.test';
-import LinkTo from './link.tsx';
+import LinkTo from './link';
 
 jest.mock('@storybook/addons');