+++ lib/components/src/bar/separator.tsx
@@ -1,25 +1,6 @@
 import React, { Fragment } from 'react';
 import { styled } from '@storybook/theming';
 
-export const interleaveSeparators = (list: any[]) =>
-  list.reduce(
-    (acc, item, index) =>
-      item ? (
-        <Fragment key={item.id || item.key || `f-${index}`}>
-          {acc}
-          {index > 0 ? <Separator key={`s-${index}`} /> : null}
-          {item.render() || item}
-        </Fragment>
-      ) : (
-        acc
-      ),
-    null
-  );
-
-interface SeparatorProps {
-  force?: boolean;
-}
-
 export const Separator = styled.span<SeparatorProps>(
   ({ theme }) => ({
     width: 1,
@@ -37,3 +18,22 @@ export const Separator = styled.span<SeparatorProps>(
         }
 );
 Separator.displayName = 'Separator';
+
+export const interleaveSeparators = (list: any[]) =>
+  list.reduce(
+    (acc, item, index) =>
+      item ? (
+        <Fragment key={item.id || item.key || `f-${index}`}>
+          {acc}
+          {/* eslint-disable-next-line react/no-array-index-key */}
+          {index > 0 ? <Separator key={`s-${index}`} /> : null}
+          {item.render() || item}
+        </Fragment>
+      ) : (
+        acc
+      ),
+    null
+  );
+interface SeparatorProps {
+  force?: boolean;
+}