+++ lib/theming/src/create.ts
@@ -51,6 +51,7 @@ const createColors = (vars: ThemeVars): Color => ({
   positive: color.positive,
   negative: color.negative,
   warning: color.warning,
+  critical: color.critical,
 
   defaultText: vars.textColor || color.darkest,
   inverseText: vars.textInverseColor || color.lightest,
@@ -139,6 +140,7 @@ export const convert = (inherit: ThemeVars = lightThemeVars): Theme => {
       positive: background.positive,
       negative: background.negative,
       warning: background.warning,
+      critical: background.critical,
     },
     typography: {
       fonts: {
@@ -172,7 +174,7 @@ export const convert = (inherit: ThemeVars = lightThemeVars): Theme => {
     brand: {
       title: brandTitle,
       url: brandUrl,
-      image: brandImage || brandTitle ? null : undefined,
+      image: brandImage || (brandTitle ? null : undefined),
     },
 
     code: createSyntax({